Implemented filter_and_modify_arts, a combat function that centralizes the copying of Arts objects from our database and modifies the copies by calling appropriate other functions, e.g., berserk_check. Filter_and_modify_arts is designed to work for both attacking (CmdAttack and CmdInterrupt) and displaying info (CmdSheet, CmdArts, CmdListAttacks, CmdCheck) and, for display purposes, produces a "baseline" copy of an art for self-comparison: e.g., if the AP cost went up overall, we'll know from comparison to the baseline object (so utilities.py doesn't need to reach directly into the Arts table to compare).
I had to rename a few things, like the ap_change attribute of the Attack class to match the ap attribute of the Arts model (because we didn't want to migrate the model and rename that instead, even if ap_change is a more accurate name and ap could be confused with the current AP of a character. I'm sure this change won't ever come back to bite us in the ass.)

Everything works well in testing EXCEPT CmdCheck when called to check interrupt chances BECAUSE CmdCheck doesn't actually use the functions in utilities.py yet as we've been meaning to refactor it for a long time. So I'm pushing this mostly-complete functionality now, and I'll refactor CmdCheck next so that it populates its tables in a more standardized way.